create separate folder for auth components
