+++ qa-admin/src/main/client/src/pages/Login.tsx
@@ -1,7 +1,7 @@
 import {FC} from "react"
 import {MainLayout} from "../components/layouts/MainLayout";
 import {ValidatedInput} from "../components/UI/ValidatedInput/ValidatedInput";
-import {LoginForm} from "../components/LoginForm";
+import {LoginForm} from "../components/auth/LoginForm";
 
 const Login: FC = () => {
     return (
